requirements.txt
requests
python-whois
dnspython
colorama
pyfiglet
tabulate
bcrypt
